+++ R/bfi_items.R
@@ -276,5 +276,4 @@
 #'   \item \code{\link[TAM]{tam.mml}} for TAM's GRM estimation procedures
 #' }
 #' @keywords datasets
-
 "bfi_items"